open_weather/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "prototype-data-pipeline"
version = "0.1.0"
description = "Data Engineering Pipeline - Extraction and Storage Prototype"
requires-python = ">=3.11"
dependencies = [
    # Data Extraction
    "dlt>=0.5.0",
    "dlt[duckdb]>=0.5.0",
    "dlt[google-sheets]>=0.5.0",
    "dlt[parquet]>=0.5.0",
    
    # Storage & Processing
    "duckdb>=0.10.0",
    "polars>=0.20.0",
    "pyiceberg>=0.5.0",
    
    # Orchestration
    "dagster>=1.7.0",
    "dagster-duckdb>=0.21.0",
    "dagster-dbt>=0.23.0",
    
    # Data Quality
    "great-expectations>=0.18.0",
    
    # Utilities
    "pyarrow>=15.0.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.3.0",
    "black>=24.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["dlt*", "dagster*"]
exclude = ["tests*", "data*", "dbt*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "integration: marks tests as integration tests (requires API key and external services)",
]
